pyproject.toml
[tool.poetry]
name = "fast-api-exercise"
version = "0.1.0"
description = ""
authors = ["humberto <matheushumberto520@gmail.com>"]
readme = "README.md"

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev fast_api_exercise/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_api_exercise -vv'
post_test = 'coverage html'

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = { extras = ["standard"], version = "^0.115.5" }
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
pyjwt = "^2.10.1"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
psycopg = {extras = ["binary"], version = "^3.2.3"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
taskipy = "^1.14.0"
ignr = "^2.2"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"
testcontainers = "^4.9.0"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
